breaks MeetingType into separate component
